refactor: split vertical table widget controls
Move vertical pricing table widget into its own folder structure
Separate content and style controls into dedicated manager and section files
Update widget manager paths to load the new widget and control directories
