+++ includes/widgets/vertical-table/controls/content-controls/content-header.php
@@ -1,0 +1,118 @@
+<?php
+
+if (! defined('ABSPATH')) {
+    exit;
+}
+
+use Elementor\Controls_Manager;
+
+trait Saasp_Vertical_Content_Header_Controls
+{
+    protected function register_content_header_controls()
+    {
+    $this->start_controls_section(
+        'saasp_vertical_content_header',
+        [
+            'label' => esc_html__( 'Header', 'saaspricing' ),
+            'tab' =>  Controls_Manager::TAB_CONTENT,
+        ]
+    );
+
+    $this->add_control(
+        'saasp_vertical_icon',
+        [
+            'label' => esc_html__( 'Icon', 'saaspricing' ),
+            'type' =>  Controls_Manager::ICONS,
+        ]
+    );
+
+    $this->add_control(
+        'saasp_vertical_header_title',
+        [
+            'label' => esc_html__( 'Title', 'saaspricing' ),
+            'type' =>  Controls_Manager::TEXT,
+            'default' => esc_html__( 'Enter Your Title', 'saaspricing' ),
+        ]
+    );
+
+    $this->add_control(
+        'saasp_vertical_header_description',
+        [
+            'label' => esc_html__( 'Description', 'saaspricing' ),
+            'type' =>  Controls_Manager::TEXT,
+            'default' => esc_html__( 'Enter Your Description', 'saaspricing' ),
+        ]
+    );
+
+    $this->add_control(
+        'saasp_vertical_header_title_tag',
+        [
+            'label' => esc_html__( 'Title HTML Tag', 'saaspricing' ),
+            'type' =>  Controls_Manager::SELECT,
+            'default' => 'h2',
+            'options' => [
+                'h2' => esc_html__( 'H2', 'saaspricing' ),
+                'h3' => esc_html__( 'H3', 'saaspricing' ),
+                'h4'  => esc_html__( 'H4', 'saaspricing' ),
+                'h5' => esc_html__( 'H5', 'saaspricing' ),
+                'h6' => esc_html__( 'H6', 'saaspricing' ),
+                'span' => esc_html__( 'Span', 'saaspricing' ),
+                'p' => esc_html__( 'P', 'saaspricing' ),
+            ],
+        ]
+    );
+    
+    $this->add_control(
+        'saasp_vertical_show_ribbon',
+        [
+            'label' => esc_html__( 'Ribbon', 'saaspricing' ),
+            'type' =>  Controls_Manager::SWITCHER,
+            'label_on' => esc_html__( 'Show', 'saaspricing' ),
+            'label_off' => esc_html__( 'Hide', 'saaspricing' ),
+            'return_value' => 'yes',
+            'default' => 'no',
+        ]
+    );
+
+    $this->add_control(
+        'saasp_vertical_ribbon_title',
+        [
+            'label' => esc_html__( 'Title', 'saaspricing' ),
+            'type' =>  Controls_Manager::TEXT,
+            'default' => esc_html__( 'Popular', 'saaspricing' ),
+            'condition' => [
+                'saasp_vertical_show_ribbon' => 'yes',
+            ],
+        ]
+    );
+
+    $this->add_control(
+        'sassp_vertical_ribbon_alignment',
+        [
+            'label' => esc_html__( 'Header Alignment', 'saaspricing' ),
+            'type' =>  Controls_Manager::CHOOSE,
+            'options' => [
+                'left' => [
+                    'title' => esc_html__( 'Left', 'saaspricing' ),
+                    'icon' => 'eicon-text-align-left',
+                ],
+                'center' => [
+                    'title' => esc_html__( 'Center', 'saaspricing' ),
+                    'icon' => 'eicon-text-align-center',
+                ],
+                'right' => [
+                    'title' => esc_html__( 'Right', 'saaspricing' ),
+                    'icon' => 'eicon-text-align-right',
+                ],
+            ],
+            'default' => 'center', 
+            'toggle' => true,
+            'selectors' => [
+                '{{WRAPPER}} .saaspricing-vertical-header-alignment' => 'text-align: {{VALUE}};',
+            ],
+        ]
+    );
+    
+    $this->end_controls_section();
+    }
+}
