+++ specs/001-split-widget-controls/plan.md
@@ -98,7 +98,15 @@ includes/
     |       |   `-- content-*.php
     |       `-- style-controls/
     |           `-- style-*.php
-    |-- saaspricing-vertical-table.php
+    |-- vertical-table/
+    |   |-- saaspricing-vertical-table.php
+    |   `-- controls/
+    |       |-- content-controls-manager.php
+    |       |-- style-controls-manager.php
+    |       |-- content-controls/
+    |       |   `-- content-*.php
+    |       `-- style-controls/
+    |           `-- style-*.php
 assets/
 |-- css/
 `-- js/
